Interface webapi para mhs con el método mi3/salidas

+++ WMS.DALCore/I_nav_ped_compra_enc/clsLnI_nav_ped_compra_enc.cs
@@ -8,7 +8,6 @@ using Microsoft.VisualBasic.CompilerServices;
 using WMS.DALCore.I_nav_barras_pallet;
 using WMS.EntityCore;
 using WMS.EntityCore.I_nav_Ped_Compra;
-using WMS.EntityCore.Interface;
 using WMS.EntityCore.Pedido;
 using WMS.EntityCore.Picking;
 using WMS.EntityCore.Producto;
@@ -1170,10 +1169,10 @@ public class clsLnI_nav_ped_compra_enc
                         throw new Exception("ERROR_202210251745: El factor es 0 para la presentación NO se puede inferir la conversión.");
                     }
 
-                    decimal v1 = Convert.ToDecimal(vCantidadEnteraPres);
-                    decimal v2 = Convert.ToDecimal(vCantidadDecimalUMBas);
+                    double v1 = Convert.ToDouble(vCantidadEnteraPres);
+                    double v2 = Convert.ToDouble(vCantidadDecimalUMBas);
 
-                    clsPublic.Split_Decimal(Convert.ToDecimal(navPedidoCompraDet.Quantity / BePresentacion.Factor),
+                    clsPublic.Split_Decimal(Convert.ToDouble(navPedidoCompraDet.Quantity / BePresentacion.Factor),
                                             ref v1,
                                             ref v2);
